services/recompute/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.9
python-dotenv==1.1.1
pydantic==2.11.7
pgvector==0.3.6
numpy==1.26.3